refac: create trait TodoFile * create module storage * create module storage::file
